raptor_xml_escape_string changed return value to int

+++ src/raptor.h
@@ -331,7 +331,7 @@ RAPTOR_API void raptor_free_identifier(raptor_identifier *identifier);
 RAPTOR_API int raptor_print_ntriples_string(FILE *stream, const unsigned char *string, const char delim);
 RAPTOR_API unsigned char* raptor_ntriples_string_as_utf8_string(raptor_parser* rdf_parser, unsigned char *src, int len, size_t *dest_lenp);
 RAPTOR_API const char* raptor_ntriples_term_as_string (raptor_ntriples_term_type term);
-RAPTOR_API size_t raptor_xml_escape_string(const unsigned char *string, size_t len, unsigned char *buffer, size_t length, char quote, raptor_simple_message_handler error_handler, void *error_data);
+RAPTOR_API int raptor_xml_escape_string(const unsigned char *string, size_t len, unsigned char *buffer, size_t length, char quote, raptor_simple_message_handler error_handler, void *error_data);
 
 RAPTOR_API char* raptor_vsnprintf(const char *message, va_list arguments);
